+++ CHANGES.rst
@@ -1,14 +1,28 @@
 CHANGES
 =======
 
-New builtins
+2.1.0
+-----
+
+  New builtins
+++++++++++++++
+
+* ``ByteArray``
+* ``FileNames``
+* ``CreateFile``
+* ``CreateTemporary``
+
+
+Enhancements
 ++++++++++++
-ByteArray
-FileNames
-CreateFile
-CreateTemporary
 
+* ``FileNameJoin`` - implement ``OperatingSystem`` option
 
+Miscellanea
++++++++++++
+
+A pass was made to improve Microsof Windows compatability and testin
+Windows under MSYS.
 
 2.0.0
 -----
@@ -88,7 +102,7 @@ Numerous bugs were fixed while working on Combinatorica V0.9 and CellsToTeX.
 Document updates
 ++++++++++++++++
 
-- Start a readthedocs `Developer Guide <https://mathics-development-guide.readthedocs.io/en/latest/>`_
+- Start a readthedocs `Developer Guide <https://mathics-development-guide.reandthedocs.io/en/latest/>`_
 
 Enhancements and bug fixes:
 +++++++++++++++++++++++++++